[login] Cleanup, split out connect_server.h, delete old files.
